swagger : swagger UI implemented

+++ src/main/java/com/demo/hotel/allocation/model/OccupancyResponse.java
@@ -1,4 +1,4 @@
 package com.demo.hotel.allocation.model;
 
-public record OccupancyResponse(int usagePremium, double revenuePremium, int usageEconomy, double revenueEconomy) {
+public record OccupancyResponse(int usagePremium, int revenuePremium, int usageEconomy, double revenueEconomy) {
 }
